Moved the Gateway files under /src directory.  Updated main.go to call new Sleeper handlers under sleeper package
